+++ packages/iov-tendermint-rpc/types/encodings.d.ts
@@ -32,6 +32,14 @@ export declare function assertNumber(value: number): number;
  */
 export declare function assertArray<T>(value: ReadonlyArray<T>): ReadonlyArray<T>;
 /**
+ * A runtime checker that ensures a given value is an object in the sense of JSON
+ * (an unordered collection of key–value pairs where the keys are strings)
+ *
+ * This is used when you want to verify that data at runtime matches the expected type.
+ * This implies assertSet.
+ */
+export declare function assertObject<T>(value: T): T;
+/**
  * Throws an error if value matches the empty value for the
  * given type (array/string of length 0, number of value 0, ...)
  *