+++ packages/iov-tendermint-rpc/types/index.d.ts
@@ -1,4 +1,5 @@
 export { v0_27 } from "./v0-27";
+export { v0_31 } from "./v0-31";
 export { Client } from "./client";
 export { AbciInfoRequest, AbciQueryParams, AbciQueryRequest, BlockRequest, BlockchainRequest, BlockResultsRequest, BroadcastTxRequest, BroadcastTxParams, CommitRequest, GenesisRequest, HealthRequest, Method, Request, QueryString, QueryTag, StatusRequest, SubscriptionEventType, TxParams, TxRequest, TxSearchParams, TxSearchRequest, ValidatorsRequest, } from "./requests";
 export * from "./responses";