+++ packages/iov-tendermint-rpc/types/v0-31/hasher.d.ts
@@ -1,0 +1,2 @@
+import { TxBytes, TxHash } from "../types";
+export declare function hashTx(tx: TxBytes): TxHash;